refactor(routing): renomme tanstackRouter.tsx en router/index.tsx
